rename nfl to nfl_games
