+++ sportsdataverse/nfl/__init__.py
@@ -2,4 +2,4 @@ from sportsdataverse.nfl.nfl_loaders import *
 from sportsdataverse.nfl.nfl_pbp import *
 from sportsdataverse.nfl.nfl_schedule import *
 from sportsdataverse.nfl.nfl_teams import *
-from sportsdataverse.nfl.nfl import *
+from sportsdataverse.nfl.nfl_games import *
